+++ blender-qmm/mats/alloys/Brass.py
@@ -58,4 +58,4 @@ class QMMBrass(bpy.types.Operator):
             links(specular_group.outputs[0], BSDF.inputs[7])
             links(specular_group.outputs[1], BSDF.inputs[16])
 
-            return {'FINISHED'}
+        return {'FINISHED'}