+++ blender-qmm/mats/base/CopperMin.py
@@ -56,4 +56,13 @@ class QMMCopperMin(bpy.types.Operator):
             links(specular_group.outputs[0], BSDF.inputs[7])
             links(specular_group.outputs[1], BSDF.inputs[16])
 
-            return {'FINISHED'}
+            #CopperColorsGroup
+            bpy.ops.node.copper_colors_group_operator()
+            nodes = m_copper_min.node_tree.nodes
+            copper_colors_group = nodes.new("ShaderNodeGroup")
+            copper_colors_group.node_tree = bpy.data.node_groups['Copper Colors']
+            copper_colors_group.location = (-500, 0)
+            links = m_copper_min.node_tree.links.new
+            links(copper_colors_group.outputs[1], BSDF.inputs[0])
+
+        return {'FINISHED'}