+++ blender-qmm/mats/base/FresnelCopper.py
@@ -85,4 +85,15 @@ class QMMCopperFresnel(bpy.types.Operator):
 
             bpy.context.object.active_material = m_copper
 
-            return {'FINISHED'}
+            #CopperColorsGroup
+            bpy.ops.node.copper_colors_group_operator()
+            nodes = m_copper.node_tree.nodes
+            copper_colors_group = nodes.new("ShaderNodeGroup")
+            copper_colors_group.node_tree = bpy.data.node_groups['Copper Colors']
+            copper_colors_group.location = (-1000, 0)
+            links = m_copper.node_tree.links.new
+            links(copper_colors_group.outputs[0], m_diffuse.inputs[0])
+            links(copper_colors_group.outputs[0], m_glossy.inputs[0])
+            links(copper_colors_group.outputs[0], m_glossy2.inputs[0])
+
+        return {'FINISHED'}