+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ### 1.3.2 <!-- 12/30/21 -->
+- Bugfix: Specular Group was being added twice
+- Added Copper Color node Group with 5 color choices <!-- https://en.wikipedia.org/wiki/Copper_(color) -->
+	- **Copper** <!-- #B87333 -->
+	- **Pale Copper** <!-- #DA8A67 -->
+	- **Copper Red** <!-- #CB6D51 -->
+	- **Copper Penny** <!-- #AD6F69 -->
+	- **Copper Rose** <!-- #996666 -->
+
+### 1.3.2 <!-- 12/29/21 -->
 - Combine IOR/Specular math nodes into Specular Group
 
 ### 1.3.1 <!-- 12/21/21 -->